trade 서비스 매수/매도 정책 및 테스트 오류 수정

+++ trade-service/src/main/java/com/moni/trade/asset/application/calculator/model/HoldingInput.java
@@ -4,7 +4,7 @@ import java.math.BigDecimal;
 
 public record HoldingInput(
         String ticker,
-        Long quantity,
+        BigDecimal quantity,
         BigDecimal averagePurchasePrice,
         BigDecimal totalPurchaseAmount
 ) {
